+++ src/nodes/tap.rs
@@ -92,10 +92,6 @@ struct TapFactory {
 }
 
 impl NodeFactory for TapFactory {
-    fn describe(&self, _io: &mut crate::graph::NodeIoBuilder<'_>) -> GraphResult<()> {
-        Ok(())
-    }
-
     fn build(
         self: Box<Self>,
         _ctx: &mut crate::graph::BuildCtx,